move spotify code generator file
