+++ __tests__/frontend.t.js
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from "@testing-library/react";
 
 // Import the component to be tested
-import HomePage from "../src/app/Index";
+import HomePage from "../src/app/page";
 
 // https://github.com/vercel/next.js/discussions/58994
 jest.mock("next/navigation", () => ({
